moved FromStr impl for Primitive to its own file
